use raw PyObject* instead of Py::Object for module reference
Trying to automate python reference counting with move semantics sounds like a good idea in theory,
but in practice because copy, moves and such are mostly implicit and invisible,
it's very easy to not see where a reference count gets incremented ot decremented.
Since the GIL must be held for each of these operations, trying to have any of this
be done automatically causes a big mess instead of helping.
Let's just use the raw CPython API, so refcount modifications become explicit.